Reorganize Safex unit tests
