+++ tests/unit_tests/safex_db/safex_blockchain_fee.cpp
@@ -42,7 +42,7 @@
 #include "blockchain_db/lmdb/db_lmdb.h"
 #include "cryptonote_basic/cryptonote_format_utils.h"
 #include "cryptonote_core/cryptonote_tx_utils.h"
-#include "safex_test_common.h"
+#include "../safex_test_common.h"
 
 using namespace cryptonote;
 using epee::string_tools::pod_to_hex;
