+++ tests/unit_tests/safex_db/simple_purchase.cpp
@@ -41,7 +41,9 @@
 #include "cryptonote_basic/cryptonote_format_utils.h"
 #include "safex/safex_account.h"
 #include "safex/safex_offer.h"
-#include "safex_test_common.h"
+#include "safex/safex_purchase.h"
+#include "safex/safex_feedback.h"
+#include "../safex_test_common.h"
 
 
 using namespace cryptonote;
@@ -62,10 +64,11 @@ namespace
 
 
   template<typename T>
-  class SafexOfferTest : public testing::Test
+  class SimplePurchaseTest : public testing::Test
   {
     protected:
-      SafexOfferTest() : m_db(new T(false, cryptonote::network_type::FAKECHAIN)), m_hardfork(*m_db, 1, 0)
+
+      SimplePurchaseTest() : m_db(new T(false, cryptonote::network_type::FAKECHAIN)), m_hardfork(*m_db, 1, 0)
       {
         m_test_sizes = std::vector<size_t>(NUMBER_OF_BLOCKS, 0);
         m_test_coins = std::vector<uint64_t>(NUMBER_OF_BLOCKS, 60);
@@ -99,13 +102,15 @@ namespace
         data1_new = std::vector<uint8_t>(data1_new_str.begin(), data1_new_str.end());
 
 
-        m_safex_offer[0] = safex::safex_offer("Apple",10,100*COIN,"This is an apple",  m_safex_account1.username,m_users_acc[0].get_keys().m_view_secret_key,m_users_acc[0].get_keys().m_account_address);
-        m_safex_offer[1] = safex::safex_offer("Barbie",30,500*COIN,"This is a Barbie", m_safex_account2.username,m_users_acc[1].get_keys().m_view_secret_key,m_users_acc[1].get_keys().m_account_address);
-        m_safex_offer[2] = safex::safex_offer("Car",1,1000*COIN,"This is a car", m_safex_account1.username,m_users_acc[0].get_keys().m_view_secret_key,m_users_acc[0].get_keys().m_account_address);
+        m_safex_offer[0] =  safex::safex_offer("Apple", 10, 10*SAFEX_CASH_COIN,"This is an apple",  m_safex_account1.username,m_users_acc[0].get_keys().m_view_secret_key,m_users_acc[0].get_keys().m_account_address);
+        m_safex_offer[1] =  safex::safex_offer("Barbie", 30, 50*SAFEX_CASH_COIN,"This is a Barbie", m_safex_account2.username,m_users_acc[0].get_keys().m_view_secret_key,m_users_acc[0].get_keys().m_account_address);
 
-        m_safex_price_peg = safex::safex_price_peg("USD price peg",m_safex_account1.username, "USD", "xcalibra USD price peg", 30);
+        m_safex_purchase = safex::safex_purchase{1, m_safex_offer[0].price, m_safex_offer[0].offer_id, true};
 
-        m_safex_offer[2].set_price_peg(m_safex_price_peg.price_peg_id,100,1000*COIN);
+        m_safex_feedback = safex::safex_feedback{3,"Eating it all day", m_safex_offer[0].offer_id};
+
+        offers_total_fee = calculate_safex_network_fee(m_safex_purchase.price, FAKECHAIN, safex::command_t::simple_purchase);
+
 
         std::string new_str_desc{"Now without worms!!"};
         std::vector<uint8_t> new_desc{new_str_desc.begin(),new_str_desc.end()};
@@ -162,13 +167,6 @@ namespace
             construct_create_account_transaction(m_txmap, m_blocks, tx2, m_users_acc[1], default_miner_fee, 0, m_safex_account2.username, m_safex_account2.pkey, m_safex_account2.account_data, m_safex_account2_keys.get_keys());
             m_txmap[get_transaction_hash(tx2)] = tx2;
           }
-          else if (i == 6)
-          {
-            tx_list.resize(tx_list.size() + 1);
-            cryptonote::transaction &tx = tx_list.back();                                                           \
-            construct_create_price_peg_transaction(m_txmap, m_blocks, tx, m_users_acc[0], default_miner_fee, 0, m_safex_account1.pkey, m_safex_price_peg, m_safex_account1_keys.get_keys());
-            m_txmap[get_transaction_hash(tx)] = tx;
-          }
           else if (i == 7)
           {
             tx_list.resize(tx_list.size() + 1);
@@ -183,25 +181,27 @@ namespace
           }
           else if (i == 12)
           {
-              tx_list.resize(tx_list.size() + 1);
-              cryptonote::transaction &tx = tx_list.back();                                                           \
-              construct_create_offer_transaction(m_txmap, m_blocks, tx, m_users_acc[0], default_miner_fee, 0, m_safex_account1.pkey, m_safex_offer[2], m_safex_account1_keys.get_keys());
-              m_txmap[get_transaction_hash(tx)] = tx;
+
           }
           else if (i == 14)
           {
-            tx_list.resize(tx_list.size() + 1);
-            cryptonote::transaction &tx = tx_list.back();                                                           \
-            construct_edit_account_transaction(m_txmap, m_blocks, tx, m_users_acc[0], default_miner_fee, 0, m_safex_account1.username, data1_new, m_safex_account1_keys.get_keys());
-            m_txmap[get_transaction_hash(tx)] = tx;
+
           }
           else if (i == 16)
           {
               tx_list.resize(tx_list.size() + 1);
               cryptonote::transaction &tx = tx_list.back();                                                           \
-              construct_edit_offer_transaction(m_txmap, m_blocks, tx, m_users_acc[0], default_miner_fee, 0, m_safex_account1.pkey, m_edited_safex_offer, m_safex_account1_keys.get_keys());
+              construct_create_purchase_transaction(m_txmap, m_blocks, tx, m_users_acc[0], default_miner_fee, 0, m_safex_purchase,m_users_acc[1].get_keys().m_account_address);
               m_txmap[get_transaction_hash(tx)] = tx;
           }
+          else if (i == 25)
+          {
+              tx_list.resize(tx_list.size() + 1);
+              cryptonote::transaction &tx = tx_list.back();                                                           \
+              construct_create_feedback_transaction(m_txmap, m_blocks, tx, m_users_acc[0], default_miner_fee, 0, m_safex_feedback);
+              m_txmap[get_transaction_hash(tx)] = tx;
+          }
+
 
           construct_block(blk, i, prev_hash, m_miner_acc, 0, m_test_sizes[i], tx_list);
 
@@ -211,7 +211,7 @@ namespace
       }
 
 
-      ~SafexOfferTest()
+      ~SimplePurchaseTest()
       {
         delete m_db;
         remove_files(m_filenames, m_prefix);
@@ -239,14 +239,17 @@ namespace
       safex::safex_account m_safex_account1;
       safex::safex_account m_safex_account2;
 
-      safex::safex_offer m_safex_offer[3];
+      safex::safex_offer m_safex_offer[2];
 
       safex::safex_offer m_edited_safex_offer;
 
-      safex::safex_price_peg m_safex_price_peg;
+      safex::safex_purchase m_safex_purchase;
 
+      safex::safex_feedback m_safex_feedback;
 
       std::vector<uint8_t> data1_new;
+
+      uint64_t offers_total_fee;
 
 
       void init_hard_fork()
@@ -275,11 +278,11 @@ namespace
 
   typedef Types<BlockchainLMDB> implementations;
 
-  TYPED_TEST_CASE(SafexOfferTest, implementations);
+  TYPED_TEST_CASE(SimplePurchaseTest, implementations);
 
 #if 1
 
-  TYPED_TEST(SafexOfferTest, CreateOfferCommand) {
+  TYPED_TEST(SimplePurchaseTest, CreatePurchaseCommand) {
         boost::filesystem::path tempPath = boost::filesystem::temp_directory_path() / boost::filesystem::unique_path();
         std::string dirPath = tempPath.string();
 
@@ -305,12 +308,6 @@ namespace
             ASSERT_TRUE(std::equal(safex_offer.description.begin(), safex_offer.description.end(),
                                    saved_offer.description.begin()));
             ASSERT_EQ(safex_offer.title,saved_offer.title);
-            ASSERT_EQ(safex_offer.seller_private_view_key, saved_offer.seller_private_view_key);
-            ASSERT_EQ(safex_offer.seller_address, saved_offer.seller_address);
-
-            ASSERT_EQ(safex_offer.price_peg_used, saved_offer.price_peg_used);
-            ASSERT_EQ(safex_offer.price_peg_id, saved_offer.price_peg_id);
-            ASSERT_EQ(safex_offer.min_sfx_price, saved_offer.min_sfx_price);
 
             std::string username;
             result = this->m_db->get_offer_seller(safex_offer.offer_id, username);
@@ -338,45 +335,30 @@ namespace
             ASSERT_NO_THROW(this->m_db->add_block(this->m_blocks[i], this->m_test_sizes[i], this->m_test_diffs[i],
                                                   this->m_test_coins[i], this->m_test_tokens[i], this->m_txs[i]));
         }
+        //Checking the quantity for purchased offer is reduced
+        safex::safex_offer purchased_offer;
+        result = this->m_db->get_offer(this->m_safex_purchase.offer_id,purchased_offer);
+        ASSERT_TRUE(result);
+        ASSERT_EQ(this->m_safex_offer[0].quantity-this->m_safex_purchase.quantity, purchased_offer.quantity);
         //Checking edited offer
         safex::safex_offer saved_offer;
-        result = this->m_db->get_offer(this->m_edited_safex_offer.offer_id,saved_offer);
-        ASSERT_TRUE(result);
-        ASSERT_TRUE(std::equal(this->m_edited_safex_offer.description.begin(), this->m_edited_safex_offer.description.end(),
-                               saved_offer.description.begin()));
-        ASSERT_EQ(this->m_edited_safex_offer.title,saved_offer.title);
-        ASSERT_EQ(this->m_edited_safex_offer.seller_private_view_key, saved_offer.seller_private_view_key);
-        ASSERT_EQ(this->m_edited_safex_offer.seller_address, saved_offer.seller_address);
-
-        ASSERT_EQ(this->m_edited_safex_offer.price_peg_used, saved_offer.price_peg_used);
-        ASSERT_EQ(this->m_edited_safex_offer.price_peg_id, saved_offer.price_peg_id);
-        ASSERT_EQ(this->m_edited_safex_offer.min_sfx_price, saved_offer.min_sfx_price);
-
-        std::string username;
-        result = this->m_db->get_offer_seller(this->m_edited_safex_offer.offer_id, username);
-        ASSERT_TRUE(result);
-        ASSERT_EQ(username.compare(this->m_edited_safex_offer.seller), 0);
-
-        uint64_t price;
-        result = this->m_db->get_offer_price(this->m_edited_safex_offer.offer_id, price);
-        ASSERT_TRUE(result);
-        ASSERT_EQ(price, this->m_edited_safex_offer.price);
-
-        uint64_t quantity;
-        result = this->m_db->get_offer_quantity(this->m_edited_safex_offer.offer_id, quantity);
-        ASSERT_TRUE(result);
-        ASSERT_EQ(this->m_edited_safex_offer.quantity, quantity);
-
-        bool active;
-        result = this->m_db->get_offer_active_status(this->m_edited_safex_offer.offer_id, active);
-        ASSERT_TRUE(result);
-        ASSERT_EQ(this->m_edited_safex_offer.active, active);
 
         for (int i = NUMBER_OF_BLOCKS2; i < NUMBER_OF_BLOCKS3; i++) {
             ASSERT_NO_THROW(this->m_db->add_block(this->m_blocks[i], this->m_test_sizes[i], this->m_test_diffs[i],
                                                   this->m_test_coins[i], this->m_test_tokens[i], this->m_txs[i]));
         }
 
+        uint64_t stars_given;
+        result = this->m_db->get_offer_stars_given(this->m_safex_purchase.offer_id, stars_given);
+        ASSERT_TRUE(result);
+        ASSERT_EQ(this->m_safex_feedback.stars_given*COIN, stars_given);
+
+        uint64_t fee_sum = 0;
+
+        for(int i=0;i<NUMBER_OF_BLOCKS3; i++) {
+            fee_sum += this->m_db->get_network_fee_sum_for_interval(i);
+        }
+        ASSERT_EQ(fee_sum,this->offers_total_fee);
 
     ASSERT_NO_THROW(this->m_db->close());
